Outsourced logic inside routes/api/profiles.js to controllers in controllers/profiles.js
